# backend/swagger/swagger.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/user/":{"get":{"tags":["user","user"],"summary":"Find Users","operationId":"find_users_user__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response Find Users User  Get"}}}}}}},"/user/{user_id}":{"get":{"tags":["user","user"],"summary":"Find User","operationId":"find_user_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/register":{"post":{"tags":["user","user"],"summary":"Create User","operationId":"create_user_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth","auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chatroom/create":{"post":{"tags":["chatroom","chatroom"],"summary":"Create Chatroom","operationId":"create_chatroom_chatroom_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatroomRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatroomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chatroom/":{"get":{"tags":["chatroom","chatroom"],"summary":"Find All Chatroom","operationId":"find_all_chatroom_chatroom__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ChatroomResponse"},"type":"array","title":"Response Find All Chatroom Chatroom  Get"}}}}}}},"/chatroom/{chatroom_id}":{"get":{"tags":["chatroom","chatroom"],"summary":"Find Chatroom","operationId":"find_chatroom_chatroom__chatroom_id__get","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatroomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["chatroom","chatroom"],"summary":"Delete Chatroom","operationId":"delete_chatroom_chatroom__chatroom_id__delete","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["chatroom","chatroom"],"summary":"Update Chatroom","operationId":"update_chatroom_chatroom__chatroom_id__patch","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChatroomRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChatroomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{chatroom_id}":{"post":{"tags":["message","message"],"summary":"Send Message To Model","operationId":"send_message_to_model_message__chatroom_id__post","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["message","message"],"summary":"Get Message","operationId":"get_message_message__chatroom_id__get","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"title":"Chatroom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetMessageResponse"},"title":"Response Get Message Message  Chatroom Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["message","message"],"summary":"Delete Message","operationId":"delete_message_message__chatroom_id__delete","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{chatroom_id}/{message_id}":{"delete":{"tags":["message","message"],"summary":"Delete Message","operationId":"delete_message_message__chatroom_id___message_id__delete","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}},{"name":"message_id","in":"path","required":true,"schema":{"type":"integer","title":"Message Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/document/{chatroom_id}":{"post":{"tags":["document","document"],"summary":"Upload Document","operationId":"upload_document_document__chatroom_id__post","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_document__chatroom_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["document","document"],"summary":"Get Documents","operationId":"get_documents_document__chatroom_id__get","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetDocumentRequest"},"title":"Response Get Documents Document  Chatroom Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/document/{chatroom_id}/{document_id}":{"delete":{"tags":["document","document"],"summary":"Delete Document","operationId":"delete_document_document__chatroom_id___document_id__delete","parameters":[{"name":"chatroom_id","in":"path","required":true,"schema":{"type":"integer","title":"Chatroom Id"}},{"name":"document_id","in":"path","required":true,"schema":{"type":"integer","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_document_document__chatroom_id__post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_document_document__chatroom_id__post"},"ChatroomResponse":{"properties":{"chatroom_id":{"type":"integer","title":"Chatroom Id"},"chatroom_name":{"type":"string","title":"Chatroom Name"},"course_code":{"type":"string","title":"Course Code"},"instructor_name":{"type":"string","title":"Instructor Name"}},"type":"object","required":["chatroom_id","chatroom_name","course_code","instructor_name"],"title":"ChatroomResponse"},"CreateChatroomRequest":{"properties":{"chatroom_name":{"type":"string","title":"Chatroom Name"},"instructor_name":{"type":"string","title":"Instructor Name"},"course_code":{"type":"string","title":"Course Code"}},"type":"object","required":["chatroom_name","instructor_name","course_code"],"title":"CreateChatroomRequest"},"CreateChatroomResponse":{"properties":{"chatroom_id":{"type":"integer","title":"Chatroom Id"}},"type":"object","required":["chatroom_id"],"title":"CreateChatroomResponse"},"GetDocumentRequest":{"properties":{"document_id":{"type":"integer","title":"Document Id"},"document_name":{"type":"string","title":"Document Name"},"uploaded_time":{"type":"string","format":"date-time","title":"Uploaded Time"}},"type":"object","required":["document_id","document_name","uploaded_time"],"title":"GetDocumentRequest"},"GetMessageResponse":{"properties":{"message_id":{"type":"integer","title":"Message Id"},"content":{"type":"string","title":"Content"},"send_time":{"type":"string","format":"date-time","title":"Send Time"},"sender_type":{"type":"string","title":"Sender Type"},"chatroom_id":{"type":"integer","title":"Chatroom Id"}},"type":"object","required":["message_id","content","send_time","sender_type","chatroom_id"],"title":"GetMessageResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SendMessageRequest":{"properties":{"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"SendMessageRequest"},"SendMessageResponse":{"properties":{"message_id":{"type":"integer","title":"Message Id"},"content":{"type":"string","title":"Content"},"send_time":{"type":"string","format":"date-time","title":"Send Time"},"sender_type":{"type":"string","title":"Sender Type"},"chatroom_id":{"type":"integer","title":"Chatroom Id"}},"type":"object","required":["message_id","content","send_time","sender_type","chatroom_id"],"title":"SendMessageResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UpdateChatroomRequest":{"properties":{"chatroom_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Chatroom Name"},"instructor_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Instructor Name"},"course_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Course Code"}},"type":"object","title":"UpdateChatroomRequest"},"UpdateChatroomResponse":{"properties":{"chatroom_id":{"type":"integer","title":"Chatroom Id"},"chatroom_name":{"type":"string","title":"Chatroom Name"},"course_code":{"type":"string","title":"Course Code"},"instructor_name":{"type":"string","title":"Instructor Name"}},"type":"object","required":["chatroom_id","chatroom_name","course_code","instructor_name"],"title":"UpdateChatroomResponse"},"UserCreateRequest":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["first_name","last_name","email","password"],"title":"UserCreateRequest"},"UserCreateResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["user_id"],"title":"UserCreateResponse"},"UserLoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLoginRequest"},"UserResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","title":"Email"}},"type":"object","required":["user_id","first_name","last_name","email"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}